+++ lab_09_01_1/user_interface.h
@@ -4,9 +4,9 @@
 
 #ifndef LAB_09_USER_INTERFACE_H
 
-int64_t read_matrix(matrix *a_m);
-int64_t print_matrix(matrix a_m);
-int64_t read_exp(int64_t *pheta, int64_t *gamma);
+int read_matrix(matrix *a_m);
+int print_matrix(matrix a_m);
+int read_exp(int *pheta, int *gamma);
 
 #define LAB_09_USER_INTERFACE_H
